pyproject.toml
[project]
name = "rfp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "openai>=1.51.2",
]
